[v2 PATCH] qemu-img: sort block formats in help message
The help message for qemu-img lists the supported block formats, of
which there are 27 as of version 2.0.50. The formats are printed in
the order of their driver's position in a linked list, which appears
random. This patch prints the formats in sorted order, making it
easier to read and to find a specific format in the list.

v2: Incorporated feedback from Stefan Hajnoczi

Signed-off-by: Mike Day <ncmike@ncultra.org>